refactor(godot): address review — rename to godot_resource, wire analyze/build/MIGRATION, README + fixtures (#1929)
- Register .gd (gdscript) and .tscn/.tres/project.godot (godot_resource) in analyze.py, build.py and extractors/MIGRATION.md, matching #1836.
- Rename the scene extractor to godot_resource (module, extract_godot_resource, registry key, wiring) per review.
- Shrink the pyproject godot comment to one line.
- README: add a godot extras row and split the godot file types into their own GDScript / Godot resources rows.
- Move the inline gdscript/godot test bodies into tests/fixtures/godot/ and rename the test file to test_godot_resource.py.
