+++ graphify/build.py
@@ -50,6 +50,8 @@ _EDGE_LANG_FAMILY: dict[str, str] = {
     ".cxx": "c", ".hh": "c", ".hxx": "c",
     ".cu": "c", ".cuh": "c", ".metal": "c", ".m": "c", ".mm": "c",
     ".rb": "rb", ".rake": "rb", ".php": "php", ".cs": "cs", ".swift": "swift", ".lua": "lua",
+    ".gd": "gdscript",
+    ".tscn": "godot_resource", ".tres": "godot_resource", ".godot": "godot_resource",
 }
