+++ graphify/extractors/__init__.py
@@ -19,7 +19,7 @@ from graphify.extractors.elixir import extract_elixir
 from graphify.extractors.fortran import extract_fortran
 from graphify.extractors.gdscript import extract_gdscript
 from graphify.extractors.go import extract_go
-from graphify.extractors.godot_scene import extract_godot_scene
+from graphify.extractors.godot_resource import extract_godot_resource
 from graphify.extractors.json_config import extract_json
 from graphify.extractors.julia import extract_julia
 from graphify.extractors.markdown import extract_markdown
@@ -49,7 +49,7 @@ LANGUAGE_EXTRACTORS: dict[str, Callable[[Path], dict]] = {
     "fortran": extract_fortran,
     "gdscript": extract_gdscript,
     "go": extract_go,
-    "godot_scene": extract_godot_scene,
+    "godot_resource": extract_godot_resource,
     "json": extract_json,
     "julia": extract_julia,
     "lazarus_form": extract_lazarus_form,
