+++ tests/fixtures/godot/gdscript/analytics.gd
@@ -1,0 +1,3 @@
+extends Node
+func track(name):
+	pass
