+++ tests/fixtures/godot/gdscript/audio.gd
@@ -1,0 +1,2 @@
+extends Node
+class_name AudioManager
