+++ tests/fixtures/godot/gdscript/player.gd
@@ -1,0 +1,3 @@
+extends Node
+func attack():
+	Analytics.track("hit")
