+++ tests/fixtures/godot/resource/scenes/Bullet.tscn
@@ -1,0 +1,3 @@
+[gd_scene format=3 uid="uid://bul"]
+
+[node name="Bullet" type="Area2D"]
