+++ tests/fixtures/godot/resource/scenes/Main.tscn
@@ -1,0 +1,11 @@
+[gd_scene load_steps=3 format=3 uid="uid://abc"]
+
+[ext_resource type="Script" path="res://scripts/enemy.gd" id="1_e"]
+[ext_resource type="PackedScene" path="res://scenes/Bullet.tscn" id="2_b"]
+
+[node name="Enemy" type="CharacterBody2D"]
+script = ExtResource("1_e")
+
+[node name="Hitbox" type="Area2D" parent="."]
+
+[connection signal="body_entered" from="Hitbox" to="." method="take_damage"]
