+++ tests/fixtures/godot/resource/scripts/enemy.gd
@@ -1,0 +1,4 @@
+class_name Enemy
+extends CharacterBody2D
+func take_damage(a):
+	pass
