+++ tests/fixtures/godot/resource/scripts/game_state.gd
@@ -1,0 +1,1 @@
+extends Node
