+++ tests/fixtures/godot/resource/weird.tscn
@@ -1,0 +1,6 @@
+[gd_scene format=3]
+
+[ext_resource type="Script" path="res://scripts/enemy.gd" id="1_e"]
+
+[node name="Odd]Name" type="Node2D"]
+script = ExtResource("1_e")
