+++ tests/test_gdscript.py
@@ -1,5 +1,5 @@
+import shutil
 import tempfile
-import textwrap
 import unittest
 from pathlib import Path
 
@@ -7,6 +7,7 @@ from graphify.extract import extract_gdscript, _make_id, _file_stem
 from graphify.extractors.gdscript import _load_gdscript_parser
 
 _HAS_GRAMMAR = _load_gdscript_parser() is not None
+_FIXTURES = Path(__file__).parent / "fixtures" / "godot" / "gdscript"
 
 
 def _rels(result):
@@ -20,42 +21,17 @@ def _edge(result, relation):
 @unittest.skipUnless(_HAS_GRAMMAR, "tree-sitter-gdscript grammar not installed")
 class TestGDScript(unittest.TestCase):
     def setUp(self):
+        # Copy the fixture project into a temp dir so res:// resolution and the
+        # per-project autoload cache anchor on a real, isolated project root.
         self.tmp = tempfile.TemporaryDirectory()
-        self.root = Path(self.tmp.name)
-        (self.root / "project.godot").write_text("config_version=5\n")
+        self.root = Path(self.tmp.name) / "proj"
+        shutil.copytree(_FIXTURES, self.root)
 
     def tearDown(self):
         self.tmp.cleanup()
 
-    def _write(self, name, body):
-        p = self.root / name
-        p.write_text(textwrap.dedent(body))
-        return p
-
     def test_class_extends_functions_signals_calls(self):
-        (self.root / "audio.gd").write_text("extends Node\nclass_name AudioManager\n")
-        p = self._write("enemy.gd", """
-            class_name Enemy
-            extends CharacterBody2D
-
-            signal died(reason)
-
-            func _ready():
-                died.connect(_on_died)
-                sprite.play("idle")
-
-            func take_damage(amount):
-                if amount > 0:
-                    emit_signal("died", "killed")
-                    die()
-
-            func die():
-                var fx = preload("res://audio.gd")
-                queue_free()
-
-            func _on_died(reason):
-                print(reason)
-        """)
+        p = self.root / "enemy.gd"
         r = extract_gdscript(p)
         rels = _rels(r)
         for expected in ("defines", "extends", "declares", "emits", "connects", "calls", "imports"):
@@ -88,19 +64,10 @@ class TestGDScript(unittest.TestCase):
 
     def test_autoload_method_call_resolves_to_script_function(self):
         # project.godot registers Analytics as an autoload -> analytics.gd
-        (self.root / "project.godot").write_text(
-            'config_version=5\n\n[autoload]\nAnalytics="*res://analytics.gd"\n'
-        )
-        analytics = self.root / "analytics.gd"
-        analytics.write_text("extends Node\nfunc track(name):\n\tpass\n")
-        caller = self._write("player.gd", """
-            extends Node
-            func attack():
-                Analytics.track("hit")
-        """)
+        caller = self.root / "player.gd"
         r = extract_gdscript(caller)
         # the call must target analytics.gd's track function node, NOT a bare anchor
-        want = _make_id(_file_stem(analytics), "track")
+        want = _make_id(_file_stem(self.root / "analytics.gd"), "track")
         resolved = [e for e in _edge(r, "calls") if e.get("context") == "Analytics"]
         self.assertTrue(resolved, "no autoload-resolved call edge emitted")
         self.assertEqual(resolved[0]["target"], want)
@@ -109,7 +76,7 @@ class TestGDScript(unittest.TestCase):
 
     def test_missing_grammar_is_graceful(self):
         # Even with a grammar present, an empty file yields just the file node.
-        p = self._write("empty.gd", "")
+        p = self.root / "empty.gd"
         r = extract_gdscript(p)
         self.assertTrue(any(n["label"] == "empty.gd" for n in r["nodes"]))
         self.assertNotIn("error", r)
